DrawApp/requirements.txt
Flask==3.1.0
torch
torchvision
Pillow